+++ index.html
@@ -7,13 +7,10 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/functions.js" defer></script>
-  <script src="js/main.js" defer></script>
+  <script src="./js/main.js" type="module" defer></script>
 </head>
-
 <body>
 <main>
-
   <!-- Фильтрация изображений от других пользователей -->
   <section class="img-filters  img-filters--inactive  container">
     <h2 class="img-filters__title  visually-hidden">Фильтр фотографий</h2>
