+++ js/main.js
@@ -1,151 +1,5 @@
-/**
- * Случайное число в заданном диапазоне включая min и max
- * @param min {number}
- * @param max {number}
- * @return {number}
- */
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
+import {createCards} from './create-cards.js';
 
-/**
- * Создаёт генератор уникальных чисел начиная со start
- * @param start {number}
- * @return {() => number}
- */
-const getUniqueInteger = (start = 0) => () => start++;
-
-
-/**
- * Создаёт генератор уникальных чисел в диапазоне [min ... max], включая min и max
- * @param min {number}
- * @param max {number}
- * @return {() => number}
- */
-const createRandomIdFromRangeGenerator = (min, max) => {
-  const previousValues = [];
-  return function () {
-    let currentValue = getRandomInteger(min, max);
-    if (previousValues.length >= (max - min + 1)) {
-      return null;
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(min, max);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-};
-
-/**
- * Возвращает случайный элемент из переданного массива
- * @param arr {array}
- * @return {*}
- */
-const getRandomElementFromArray = (arr) => arr[getRandomInteger(0, arr.length - 1)];
-
-//------------------------------------- TODO: перенести всё, что выше, в functions.js
-
-/**
- * @type array <string>
- */
-const COMMENT_MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-/**
- * @type array <string>
- */
-const USER_NAMES = [
-  'Артём',
-  'Егор',
-  'Клавдий Пятый',
-  'Сын маминой подруги',
-  'Стивен Спилберг',
-  'Джеф Безос',
-  'Барбара Лисков'
-];
-
-/**
- * @type array <string>
- */
-const CARD_DESCRIPTIONS = [
-  'Также как синтетическое тестирование влечет за собой процесс внедрения и модернизации вывода текущих активов. Являясь всего лишь частью общей картины, независимые государства представляют собой не что иное, как квинтэссенцию победы маркетинга над разумом и должны быть обнародованы. Есть над чем задуматься: сторонники тоталитаризма в науке могут быть разоблачены.',
-  'Идейные соображения высшего порядка, а также новая модель организационной деятельности предоставляет широкие возможности для новых предложений. Значимость этих проблем настолько очевидна, что курс на социально-ориентированный национальный проект прекрасно подходит для реализации поставленных обществом задач. Мы вынуждены отталкиваться от того, что укрепление и развитие внутренней структуры предоставляет широкие возможности для своевременного выполнения сверхзадачи.',
-  'Разнообразный и богатый опыт говорит нам, что выбранный нами инновационный путь предоставляет широкие возможности для экономической целесообразности принимаемых решений! Таким образом, высокотехнологичная концепция общественного уклада требует от нас анализа как самодостаточных, так и внешне зависимых концептуальных решений. Являясь всего лишь частью общей картины, сделанные на базе интернет-аналитики выводы, вне зависимости от их уровня, должны быть превращены в посмешище, хотя само их существование приносит несомненную пользу обществу.',
-  'Но диаграммы связей, которые представляют собой яркий пример континентально-европейского типа политической культуры, будут объединены в целые кластеры себе подобных. Принимая во внимание показатели успешности, курс на социально-ориентированный национальный проект требует анализа стандартных подходов. А ещё сторонники тоталитаризма в науке описаны максимально подробно.',
-  'Каждый из нас понимает очевидную вещь: новая модель организационной деятельности требует определения и уточнения переосмысления внешнеэкономических политик. А также ключевые особенности структуры проекта заблокированы в рамках своих собственных рациональных ограничений. Повседневная практика показывает, что курс на социально-ориентированный национальный проект не оставляет шанса для глубокомысленных рассуждений.',
-  'Таким образом, глубокий уровень погружения, в своём классическом представлении, допускает внедрение поэтапного и последовательного развития общества. Высокий уровень вовлечения представителей целевой аудитории является четким доказательством простого факта: базовый вектор развития прекрасно подходит для реализации инновационных методов управления процессами. В своём стремлении улучшить пользовательский опыт мы упускаем, что непосредственные участники технического прогресса могут быть описаны максимально подробно.',
-  'В своём стремлении улучшить пользовательский опыт мы упускаем, что многие известные личности, превозмогая сложившуюся непростую экономическую ситуацию, объявлены нарушающими общечеловеческие нормы этики и морали. Задача организации, в особенности же повышение уровня гражданского сознания напрямую зависит от дальнейших направлений развития! Есть над чем задуматься: базовые сценарии поведения пользователей, инициированные исключительно синтетически, обнародованы.'
-];
-
-//------------------------------------- TODO: перенести всё, что выше, в некий Store
-
-/** @description Типы*/
-/**
- * @typedef {object} card
- * @property {number} id,
- * @property {string} url,
- * @property {string} description,
- * @property {number} likes,
- * @property {array <comment>=} comments
- */
-
-/**
- * @typedef {object} comment
- * @property {number} id,
- * @property {string} avatar,
- * @property {string} message,
- * @property {string} name
- */
-
-// Объявление генераторов id
-const generateCardId = createRandomIdFromRangeGenerator(1, 25);
-const generateCommentId = getUniqueInteger();
-
-/**
- * Создаёт объект комментария
- * @return {comment}
- */
-const createComment = () => {
-  const commentId = generateCommentId();
-  const commentMessage = `${getRandomElementFromArray(COMMENT_MESSAGES)} ${getRandomInteger(0, 1) ? `\n${getRandomElementFromArray(COMMENT_MESSAGES)}` : ''}`;
-  return {
-    id: commentId,
-    avatar: `img/avatar-${getRandomInteger(1, 6)}}.svg`,
-    message: commentMessage,
-    name: getRandomElementFromArray(USER_NAMES),
-  };
-};
-
-/**
- * Создаёт объект карточки с комментариями
- * @return {card}
- */
-const createCard = () => {
-  const cardId = generateCardId();
-  const commentsArray = Array.from({length: getRandomInteger(0, 30)}, createComment);
-  return {
-    id: cardId,
-    url: `photos/${cardId}.jpg`,
-    description: getRandomElementFromArray(CARD_DESCRIPTIONS),
-    likes: getRandomInteger(15, 200),
-    comments: commentsArray,
-  };
-};
-/* eslint-disable */ //TODO: Убрать
-/**
- * Массив со всеми карточками
- * @type {array<card>}
- */
-const cardsArray = Array.from({length: 25}, createCard);
-console.log(cardsArray);
-/* eslint-enable */ //TODO: Убрать
+/*eslint-disable*/ // TODO: Удалить
+console.log(createCards(25));
+/*eslint-enable*/ // TODO: Удалить
