+++ js/mocks.js
@@ -1,0 +1,39 @@
+/**
+ * @type array <string>
+ */
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+/**
+ * @type array <string>
+ */
+const USER_NAMES = [
+  'Артём',
+  'Егор',
+  'Клавдий Пятый',
+  'Сын маминой подруги',
+  'Стивен Спилберг',
+  'Джеф Безос',
+  'Барбара Лисков'
+];
+
+/**
+ * @type array <string>
+ */
+const CARD_DESCRIPTIONS = [
+  'Также как синтетическое тестирование влечет за собой процесс внедрения и модернизации вывода текущих активов. Являясь всего лишь частью общей картины, независимые государства представляют собой не что иное, как квинтэссенцию победы маркетинга над разумом и должны быть обнародованы. Есть над чем задуматься: сторонники тоталитаризма в науке могут быть разоблачены.',
+  'Идейные соображения высшего порядка, а также новая модель организационной деятельности предоставляет широкие возможности для новых предложений. Значимость этих проблем настолько очевидна, что курс на социально-ориентированный национальный проект прекрасно подходит для реализации поставленных обществом задач. Мы вынуждены отталкиваться от того, что укрепление и развитие внутренней структуры предоставляет широкие возможности для своевременного выполнения сверхзадачи.',
+  'Разнообразный и богатый опыт говорит нам, что выбранный нами инновационный путь предоставляет широкие возможности для экономической целесообразности принимаемых решений! Таким образом, высокотехнологичная концепция общественного уклада требует от нас анализа как самодостаточных, так и внешне зависимых концептуальных решений. Являясь всего лишь частью общей картины, сделанные на базе интернет-аналитики выводы, вне зависимости от их уровня, должны быть превращены в посмешище, хотя само их существование приносит несомненную пользу обществу.',
+  'Но диаграммы связей, которые представляют собой яркий пример континентально-европейского типа политической культуры, будут объединены в целые кластеры себе подобных. Принимая во внимание показатели успешности, курс на социально-ориентированный национальный проект требует анализа стандартных подходов. А ещё сторонники тоталитаризма в науке описаны максимально подробно.',
+  'Каждый из нас понимает очевидную вещь: новая модель организационной деятельности требует определения и уточнения переосмысления внешнеэкономических политик. А также ключевые особенности структуры проекта заблокированы в рамках своих собственных рациональных ограничений. Повседневная практика показывает, что курс на социально-ориентированный национальный проект не оставляет шанса для глубокомысленных рассуждений.',
+  'Таким образом, глубокий уровень погружения, в своём классическом представлении, допускает внедрение поэтапного и последовательного развития общества. Высокий уровень вовлечения представителей целевой аудитории является четким доказательством простого факта: базовый вектор развития прекрасно подходит для реализации инновационных методов управления процессами. В своём стремлении улучшить пользовательский опыт мы упускаем, что непосредственные участники технического прогресса могут быть описаны максимально подробно.',
+  'В своём стремлении улучшить пользовательский опыт мы упускаем, что многие известные личности, превозмогая сложившуюся непростую экономическую ситуацию, объявлены нарушающими общечеловеческие нормы этики и морали. Задача организации, в особенности же повышение уровня гражданского сознания напрямую зависит от дальнейших направлений развития! Есть над чем задуматься: базовые сценарии поведения пользователей, инициированные исключительно синтетически, обнародованы.'
+];
+
+export const getData = () => ({COMMENT_MESSAGES, USER_NAMES, CARD_DESCRIPTIONS});
